Remove MatchID class and add public getters to comply with Firebase requirements.
Our application does not need a local implementation of the Match ID and generating an unique one to be stored on the db is a hard task. Instead when match objects are pushed to the db, this
wil automatically assign a unique id which we can obtain (and use for live updates) when we request the list of Matches from the database.